segmentation_agent/requirements.txt
Flask==3.0.0
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.4.0
joblib==1.3.2
openpyxl==3.1.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
python-dotenv==1.0.0
